modules/soar/functions/soar_handler/requirements.txt
google-cloud-logging==3.8.0
google-cloud-compute==1.15.0
sqlalchemy==2.0.23
pymysql==1.1.0
